Refactor objects.go
The objects in objects.go implements methods that can be used in other controllers. While starting the implementation of the enoexec lifecycle, this commit creates a more generic model that can be extended by controller-specific ones.
